split out exception handler
